+++ laktory/models/compute/__init__.py
@@ -1,0 +1,1 @@
+from laktory.models.compute.initscript import InitScript